Slightly refactor, and change name of useOnClickOutside.
